requirements.txt
alembic==1.18.1
bcrypt==5.0.0
email-validator==2.3.0
fastapi==0.128.0
PyJWT==2.10.1
pydantic==2.12.5
pydantic-settings==2.12.0
sqlalchemy==2.0.46
structlog==25.5.0
uvicorn==0.40.0
